Revert "Simplified typedef"
This reverts commit fecfaacbaf78e0455f08e1c213e024228846a561.

Although this typedef is still useless, it's useful to check the sanity of
the names of template parameters of the periodic traits classes. If some
template paremeters are named like typedefs of the kernel (which is valid code,
but not ideal) some compilers will bug and fail to compile.

See also commit 61ed22558cf5fdc8b66d8d81d777ca949066bf7d