Replaced 'assert' with 'continue'
This function takes a vertex handle vh and computes the coordinates
in the hypothetical situation of vh being removed and querying at vh->point().

If vh is adjacent to the infinite vertex, then vh->point() will be outside
the convex hull (most likely, could be on the boundary too). The base function
taking a point can handle that case, it returns: (empty, 0, false). Rather
than crashing, let the other function return 'false'.