Motorcycle Graphs Software version 7.0
Main changes:
- Reworked the point/location data structure: instead of adding as
little as possible (i.e. if a point is on a border, only the locations
strictly in the faces where it is useful were added), a point on
the border will have all its locations put in the data structure.

- Added temporal and geometric snapping: out of all the collisions
we can detect, only two cases are points that we have never seen
before. In those cases, the new point/times are first compared
to existing times and points to find if a close-enough entity
already exists. If it does, we use that entity instead of
creating a new point.

- A bunch of various bug fixes.