Get rid of 'corners_incidence_map_'
'corners_incidence_map_' was a map<point, surface_patch_index>.
'corners_indidences_'    is a  map<Corner_index, surface_patch_index>.

The first map was only used in Lipschitz_sizing, at a place where
we are given a point and a MeshDomain::Index.

The first map was filled in 'insert_edge()', which calls
'compute_corners_incidences()' (which fills 'corners_incidences').
Thus, we can simply grab the Corner_index and use the second map
immediately.