XHTML problem regarding TOC definition
Form the doxygen documentation:
Normally the contents between \htmlonly and \endhtmlonly is inserted as-is. When you want to insert a HTML
fragment that has block scope like a table or list which should appear outside <p>..</p>, this can lead to invalid
HTML. You can use \htmlonly[block] to make doxygen end the current paragraph and restart it after \endhtmlonly.