Sneak a new feature in this branch: detect unused .cpp files
For the moment, the feature is far from being ready:

- in case of header-only, the .cpp sources files of CGAL libraries are
  reported,

- when a dependency for a test/examples is missing, the corresponding
  .cpp files is reported,

- and I have no clue which CMake version is required.

But I found real issues.

By default, the feature is OFF. It will only be activated if the CMake
variable or cache variable `CGAL_CHECK_UNUSED_CPP_FILES` is true. We
might activate it on a per-directory basis... or have it by default,
and deactivate it per-directory. We'll see. For the moment, let's keep
it deactivated.