Replace the planes by a vector+point
The idea with those planes is a sort of pre-computation of minors of
the determinant of the orientation of `(p,q,r,s)`, with `(p,q,r)`
fixed. But the three minors are only the coordinates `(a,b,c)` of the
plane defined by `(p,q,r)`, and the coordinate `d` is the determinant
of the orientation of `(p,q,r,O)` where `O` is the origin of the
Euclidean plane. We do not want to compute that `d`!

So, instead of computing planes, one computes only the normal
`(a,b,c)` of the plan `(p,q,r)`, and a stores it with `p`. That allows
to compute the determinant of `orientation(p,q,r,s)` once `s` is
known.