Problem with single quote in code span.
In case we have a single quote in a code span it is better to have the start and end backtics of the code span as double backticks.
Sometimes the constructs lead to warnings or to garbled code. This happens in the newer versions of doxygen where some new heuristics are used to recognize texts line "it's".
See also "Warning in case of usage of a single quote in a code span." (https://github.com/doxygen/doxygen/pull/7209).