Clean weird and useless converter shenanigans in OpenMesh-based simplification
Profile::Point is by definition equal to Profile::VPM::value_type,
so it is not converting anything.

Users are free to pass a custom VPM, and in that case things will break
because there is no proper geom traits being passed, but this is an
independent issue that must be fixed.