use Triangulation_3 explicitly to make sure the base class is used
before that, one could give a DT3 or a RT3 to the remeshing function.
Remeshing would then make the triangulation invalid with respect to the
Delaunay or Regular criterion, though the TDS would remain valid.

Now it is clear that we do not expect nor return anything but a valid T3.