Never allow degenerate faces to be created in smoothing
Regardless of whether "safe moves only" is being used or not.

Degenerate faces are annoying because the normal is undefined (null_vector)
and then things get tricky.

The hope is that if a move would create a degenerate face, we can
just not move it, and hope that the degenerate face would not appear
on the next move.