Fix a warning in SDG2/SDG_Linf_2
The call to source() (resp. target()) is a call to compute_source()
(resp. compute_target()) which cofuses the compiler because it does
not know that the site being considered here is known to be
an input site, for which these computations involving uninitiliazed
data cannot take place.

The functions source/target_of_supporting_site() directly return
what source()/target() return in the case of an input site.

For some reason the compiler doesn't complain about point(), which
should do the same. One could use source_of_supporting_site
since it returns p_[0] (which is what we want when we call
site.point() with an input site), but it would be confusing,
for readers this time.