Confusing line numbers in case of warnings
Based on the problems with the line numbers in some warnings, see review of https://github.com/CGAL/cgal/pull/6688 one of the problems was the `ALIASES` for figures, here we have the code `\latexonly ^^ \endlatexonly` though the code in the `\latexonly` part should be LaTeX commands. Here the "attempted" newline is not necessary at all so the `^^` can be removed leaving an empty `\latexonly` part that can be removed as well.

There are a number of other numbering problems found in https://github.com/CGAL/cgal/pull/6688 but they are on the doxygen side as a proposed pull request https://github.com/doxygen/doxygen/pull/9465 has been submitted.