Documentation error in Manual
The construct:
```
In the previous section, we wrote that "Any model of the CGAL concept
Kernel provides what is required by the concept `ConvexHullTraits_2`.
```
leads to the fact that the entire paragraph is badly formatted as the double quote is not properly closed.
Closing the double quote would lead to the fact that the `` `ConvexHullTraits_2` `` is not resolved properly

File:
- doc_output/Manual/tutorial_hello_world.html#intro_concept