Change criterion used in "should_flip"
The criterion that takes the scalar_product of the cross products
is maybe adapted to minimize the curvature when triangulating faces,
but should_flip() is used in PMP::isotropic_remeshing, (soon) PMP::refine(),
and PMP::remove_almost_degenerate_faces(). These algorithms aim to
produce well-shaped elements. The criterion is not adapted to these
algorithms: for example, on a flat mesh the scalar product is meaningless
so it will pick the diagonal which maximizes the product of the lengths
and product of sines, but this might create very anisotropic elements
since the sine of obtuse angles is still positive.

The "new" criterion is simply the criterion used in mesh smoothing
and the typical Delaunay criterion for surfaces.