WIP: trying a different should_flip criterion for cap flipping
The reasoning is that in this function, we are maybe not so interested
in having nicely shaped elements as much as we are trying to remove
the worst elements.

Unfortunately, the current criterion
(see https://github.com/CGAL/cgal/pull/7125)
can prevent some edge flips that would resolve a cap because
sometimes this Delaunay-like criterion can prefer a cap if it is
balanced by a very small angle on the other side. Then you have
alpha + beta < 180 and you're stuck with the cap, but maybe you
would have liked to flip still because the largest angles created
by a flip would be e.g. 120 and 120, which is not that bad.

So this PR tries to improve the worst cap angle instead by checking
only (and greedily) if we do not create a worse cap angle when we flip.

This produces better results for the cases that were problematic with
alpha + beta < 180