Fix a warning
/mnt/testsuite/include/CGAL/IO/PLY/PLY_writer.h:296:30: warning:
loop variable 'v' of type 'const unsigned int&' binds to a temporary
constructed from type 'const long unsigned int' [-Wrange-loop-construct]
  296 |       for(const ElementType& v : vec)

This is because ElementType is not the value type of the vector,
but the data type that we write, and for 64 bits integer,
we convert to 32 bits. So in that case, we can't cast yet.